fix
git-svn-id: http://svn.getalime.ru/rails_plugins/attacheable@53 53355a66-a8c0-4aef-80d1-10870d0d3e9b

+++ lib/attacheable.rb
@@ -73,7 +73,7 @@ module Attacheable
     def regenerate_thumbnails!(thumbnail = nil)
       connection.select_values("select id from #{table_name}").each do |object_id|
         object = find_by_id(object_id)
-        if object
+        if object && object.filename
           if thumbnail
             FileUtils.rm_f(object.full_filename_without_creation(thumbnail))
           else
